feat: extract lib.rs, add proxy binaries, fix MCP run dry-run diff
- Extract application logic from main.rs into src/lib.rs; main.rs is
  now a thin shim calling ast_bro::run(). This enables the new proxy
  binaries to reuse the same entry point.

- Add src/bin/ast-outline.rs and src/bin/sb.rs proxy binaries that
  delegate to ast_bro::run(), giving users three invocation names.

- Remove cli-typescript/bin/ast-outline.js (alias is now handled by
  the Rust proxy binary instead of a separate JS wrapper).

- Fix MCP run_run() dry-run rewrite: previously, when rewrite was set
  but write=false, the function only ran search() and collected match
  entries — it never called rewrite() to produce the new source, so
  no diff was shown. Now the search-only and rewrite paths are split;
  the rewrite path calls crate::run::rewrite() and, in dry-run mode,
  returns a unified diff via the newly-exposed unified_diff() helper.

- Refactor show_diff() into pub fn unified_diff() (returns String)
  + fn show_diff() (thin eprint! wrapper). CLI and MCP both use
  unified_diff(); CLI continues to print to stderr via show_diff().
