+++ cli-python/pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "ast-bro"
-version = "2.1.0"
+version = "2.2.0"
 description = "Fast, AST-based code-navigation toolkit — downloads the Rust binary on install"
 readme = "README.md"
 license = "MIT"
