+++ cli-typescript/bin/install.js
@@ -7,7 +7,7 @@ const fs = require("fs");
 const os = require("os");
 const path = require("path");
 
-const VERSION = "2.1.0";
+const VERSION = "2.2.0";
 const GITHUB_REPO = "aeroxy/ast-bro";
 const BINARY_NAME = "ast-bro";
