+++ cli-typescript/package.json
@@ -1,6 +1,6 @@
 {
   "name": "ast-bro",
-  "version": "2.1.0",
+  "version": "2.2.0",
   "publishConfig": {
     "access": "public"
   },
